requirements.txt
streamlit
numpy
pandas
scikit-learn
Pillow
tensorflow
joblib
 
 